Update 0008.String to Integer (atoi).java

+++ LeetCode/0008.String to Integer (atoi).java
@@ -29,6 +29,6 @@ class Solution {
             idx++;
         }
         
-        return sign * (int)res;
+        return sign * res;
     }
 }
